재구성, 추가: utils.js로 파일명 변경, 날짜 포멧 함수 추가, 주석 추가
